Make a generic component containing shared multi select logic
