+++ public/locales/en/multi_select_combobox.json
@@ -1,0 +1,5 @@
+{
+  "toggle_button_aria_label": "EN: Avaa valikko",
+  "selected_item_remove_button_aria_label": "EN: Poista {value}",
+  "clear_button_aria_label": "EN: Poista kaikki valinnat"
+}
