+++ public/locales/fi/administrative_division_dropdown.json
@@ -1,7 +1,4 @@
 {
   "label": "Paikka",
-  "placeholder": "Valitse Paikka",
-  "toggle_button_aria_label": "Avaa valikko",
-  "selected_item_remove_button_aria_label": "Poista {value}",
-  "clear_button_aria_label": "Poista kaikki valinnat"
+  "placeholder": "Valitse Paikka"
 }
