+++ public/locales/fi/multi_select_combobox.json
@@ -1,0 +1,5 @@
+{
+  "toggle_button_aria_label": "Avaa valikko",
+  "selected_item_remove_button_aria_label": "Poista {value}",
+  "clear_button_aria_label": "Poista kaikki valinnat"
+}
