+++ public/locales/sv/administrative_division_dropdown.json
@@ -1,7 +1,4 @@
 {
   "label": "SV: Paikka",
-  "placeholder": "SV: Valitse Paikka",
-  "toggle_button_aria_label": "SV: Avaa valikko",
-  "selected_item_remove_button_aria_label": "SV: Poista {value}",
-  "clear_button_aria_label": "SV: Poista kaikki valinnat"
+  "placeholder": "SV: Valitse Paikka"
 }
