+++ public/locales/sv/multi_select_combobox.json
@@ -1,0 +1,5 @@
+{
+  "toggle_button_aria_label": "SV: Avaa valikko",
+  "selected_item_remove_button_aria_label": "SV: Poista {value}",
+  "clear_button_aria_label": "SV: Poista kaikki valinnat"
+}
