+++ src/pages/search/map.tsx
@@ -151,6 +151,7 @@ export async function getStaticProps(context: GetStaticPropsContext) {
         "search_page_search_form",
         "map_view",
         "administrative_division_dropdown",
+        "multi_select_combobox",
       ])),
     },
     revalidate: 10,
